+++ src/gluon/packages/CreateApplication.ts
@@ -6,7 +6,7 @@ import {
     logger,
     MappedParameter,
     MappedParameters,
-    Parameter,
+    Parameter, success,
 } from "@atomist/automation-client";
 import {BitBucketServerRepoRef} from "@atomist/automation-client/operations/common/BitBucketServerRepoRef";
 import {GitCommandGitProject} from "@atomist/automation-client/project/git/GitCommandGitProject";
@@ -82,41 +82,35 @@ export class CreateApplication implements HandleCommand<HandlerResult> {
             return this.requestUnsetParameters(ctx);
         }
         // get memberId for createdBy
-        return gluonMemberFromScreenName(ctx, this.screenName)
-            .then(member => {
+        return ctx.messageClient.addressChannels({
+            text: "🚀 Your new library is being created...",
+        }, this.teamChannel).then(() => {
+                return gluonMemberFromScreenName(ctx, this.screenName)
+                    .then(member => {
 
-                // get project by project name
-                // TODO this should be a drop down for the member to select projects
-                // that he is associated with via Teams
-                return gluonProjectFromProjectName(ctx, this.projectName)
-                    .then(project => {
-                        // update project by creating new Bitbucket project (new domain concept)
-                        return axios.post(`${QMConfig.subatomic.gluon.baseUrl}/applications`,
-                            {
-                                name: this.name,
-                                description: this.description,
-                                applicationType: ApplicationType.DEPLOYABLE,
-                                projectId: project.projectId,
-                                createdBy: member.memberId,
-                            })
-                            .then(application => {
-                                return axios.put(application.headers.location,
+                        // get project by project name
+                        // TODO this should be a drop down for the member to select projects
+                        // that he is associated with via Teams
+                        return gluonProjectFromProjectName(ctx, this.projectName)
+                            .then(project => {
+                                // update project by creating new Bitbucket project (new domain concept)
+                                return axios.post(`${QMConfig.subatomic.gluon.baseUrl}/applications`,
                                     {
+                                        name: this.name,
+                                        description: this.description,
+                                        applicationType: ApplicationType.DEPLOYABLE,
                                         projectId: project.projectId,
+                                        createdBy: member.memberId,
                                         bitbucketRepository: {
                                             name: this.bitbucketRepositoryName,
                                             repoUrl: this.bitbucketRepositoryRepoUrl,
                                         },
-                                        createdBy: member.memberId,
+                                        requestConfiguration: true,
                                     });
                             });
                     });
-            })
-            .then(() => {
-                return ctx.messageClient.addressChannels({
-                    text: "🚀 Your new application is being provisioned...",
-                }, this.teamChannel);
-            });
+            },
+        ).then(success);
     }
 
     private requestUnsetParameters(ctx: HandlerContext): Promise<HandlerResult> {
@@ -191,13 +185,18 @@ export class LinkExistingApplication implements HandleCommand<HandlerResult> {
 
         logger.debug(`Linking to Gluon project: ${this.projectName}`);
 
-        return this.linkApplicationForGluonProject(ctx,
-            this.screenName,
-            this.teamChannel,
-            this.name,
-            this.description,
-            this.bitbucketRepositorySlug,
-            this.projectName);
+        return ctx.messageClient.addressChannels({
+            text: "🚀 Your new application is being created...",
+        }, this.teamChannel).then(() => {
+                return this.linkApplicationForGluonProject(ctx,
+                    this.screenName,
+                    this.name,
+                    this.description,
+                    this.bitbucketRepositorySlug,
+                    this.projectName);
+            },
+        );
+
     }
 
     private requestUnsetParameters(ctx: HandlerContext): Promise<HandlerResult> {
@@ -255,7 +254,6 @@ export class LinkExistingApplication implements HandleCommand<HandlerResult> {
 
     private linkApplicationForGluonProject(ctx: HandlerContext,
                                            slackScreeName: string,
-                                           teamSlackChannel: string,
                                            applicationName: string,
                                            applicationDescription: string,
                                            bitbucketRepositorySlug: string,
@@ -266,7 +264,6 @@ export class LinkExistingApplication implements HandleCommand<HandlerResult> {
 
                 return this.linkBitbucketRepository(ctx,
                     slackScreeName,
-                    teamSlackChannel,
                     applicationName,
                     applicationDescription,
                     bitbucketRepositorySlug,
@@ -277,7 +274,6 @@ export class LinkExistingApplication implements HandleCommand<HandlerResult> {
 
     private linkBitbucketRepository(ctx: HandlerContext,
                                     slackScreeName: string,
-                                    teamSlackChannel: string,
                                     applicationName: string,
                                     applicationDescription: string,
                                     bitbucketRepositorySlug: string,
@@ -325,6 +321,9 @@ export class LinkExistingApplication implements HandleCommand<HandlerResult> {
                     .then(() => {
                         return gluonMemberFromScreenName(ctx, slackScreeName)
                             .then(member => {
+                                const remoteUrl = _.find(repo.links.clone, clone => {
+                                    return (clone as any).name === "ssh";
+                                }) as any;
                                 return axios.post(`${QMConfig.subatomic.gluon.baseUrl}/applications`,
                                     {
                                         name: applicationName,
@@ -332,31 +331,17 @@ export class LinkExistingApplication implements HandleCommand<HandlerResult> {
                                         applicationType: ApplicationType.DEPLOYABLE,
                                         projectId: gluonProjectId,
                                         createdBy: member.memberId,
-                                    })
-                                    .then(application => {
-                                        const remoteUrl = _.find(repo.links.clone, clone => {
-                                            return (clone as any).name === "ssh";
-                                        }) as any;
-
-                                        return axios.put(application.headers.location,
-                                            {
-                                                projectId: gluonProjectId,
-                                                bitbucketRepository: {
-                                                    bitbucketId: repo.id,
-                                                    name: repo.name,
-                                                    slug: bitbucketRepositorySlug,
-                                                    remoteUrl: remoteUrl.href,
-                                                    repoUrl: repo.links.self[0].href,
-                                                },
-                                                createdBy: member.memberId,
-                                            });
+                                        bitbucketRepository: {
+                                            bitbucketId: repo.id,
+                                            name: repo.name,
+                                            slug: bitbucketRepositorySlug,
+                                            remoteUrl: remoteUrl.href,
+                                            repoUrl: repo.links.self[0].href,
+                                        },
+                                        requestConfiguration: true,
                                     });
                             })
-                            .then(() => {
-                                return ctx.messageClient.addressChannels({
-                                    text: "🚀 Your new application is being provisioned...",
-                                }, teamSlackChannel);
-                            });
+                            .then(success);
                     });
             });
     }
